ClientService/requirements.txt
flask
flask-restx
flask-cors
sqlalchemy
pymysql
py-eureka-client
pyjwt

